recipe/tests/test_tag_api.py
from unittest import TestCase
from django.contrib.auth import get_user_model
from django.urls import reverse
from rest_framework import status
from rest_framework.test import APIClient
from core.models import Tag
from recipe.srializers import TagSerializers