requirements.txt
Django==2.2.9
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
Pillow==7.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
virtualenv==16.7.9
